move PackingInfo to de.metas.handlingunits.shipping
