+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/shipping/AddTrackingInfosForInOutWithoutHUReq.java
@@ -22,7 +22,6 @@
 
 package de.metas.handlingunits.shipping;
 
-import de.metas.handlingunits.shipmentschedule.spi.impl.PackageInfo;
 import de.metas.inout.InOutId;
 import de.metas.shipping.ShipperId;
 import lombok.Builder;
