+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/shipping/CreatePackagesForInOutRequest.java
@@ -22,7 +22,6 @@
 
 package de.metas.handlingunits.shipping;
 
-import de.metas.handlingunits.shipmentschedule.spi.impl.PackageInfo;
 import de.metas.inout.InOutId;
 import de.metas.inout.model.I_M_InOut;
 import de.metas.shipping.model.ShipperTransportationId;
